backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
spotipy>=2.23.0
umap-learn>=0.5.6
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
starlette>=0.36.0
python-multipart>=0.0.9
httpx>=0.26.0
anthropic>=0.25.0
